refactor - extract NumberExtractor interface
- Custom 구분자의 존재 여부에 따라 String Calculator가 다른 NumberExtractor에 의존하도록 했습니다.
